backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
python-multipart>=0.0.6

# AI/ML
google-generativeai>=0.3.2

# HTTP
httpx>=0.26.0
aiohttp>=3.9.1

# Async
aiofiles>=25.1.0

# Media Processing
Pillow>=10.2.0
pytesseract>=0.3.10

# Platform APIs
praw>=7.7.1

# Caching (optional for Redis)
redis>=5.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.3
